More robust check for git folder
I'm now checking for common git files in a folder to determine if it is
a git folder instead of it's name.

This should work for normal git folders containing a hidden .git folder
as a sub folder.

This should also work for bare git repositories regardless of their
name (as long as the bare repo wasn't named "./.git" as that is the name
used by a standard git subfolder).

I've also moved the checking if a folder is named .git to it's own
function because I didn't like how long it was.